# crc/api.yml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: Workflow Microservice
  license:
    name: MIT
servers:
  - url: http://localhost:5000/v1.0
security:
  - jwt: ['secret']
paths:
  /sso_backdoor:
    get:
      operationId: crc.api.user.backdoor
      summary: A backdoor that allows someone to log in as a specific user, if they
        are in a staging environment.
      security: []  # Disable security for this endpoint only.
      parameters:
        - name: uid
          in: query
          required: true
          schema:
            type: string
        - name: email_address
          in: query
          required: false
          schema:
            type: string
        - name: display_name
          in: query
          required: false
          schema:
            type: string
        - name: affiliation
          in: query
          required: false
          schema:
            type: string
        - name: eppn
          in: query
          required: false
          schema:
            type: string
        - name: first_name
          in: query
          required: false
          schema:
            type: string
        - name: last_name
          in: query
          required: false
          schema:
            type: string
        - name: title
          in: query
          required: false
          schema:
            type: string
        - name: redirect_url
          in: query
          required: false
          schema:
            type: string
      tags:
        - Users
      responses:
        '304':
          description: Redirection to the hosted frontend with an auth_token header.
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: crc.api.user.verify_token
...